+++ FoundationKit/indexpath.h
@@ -37,22 +37,22 @@ class IndexPath
 {
 public:
     IndexPath();
-    IndexPath(const int index);
-    int& operator[] (const size_t index);
-    const int& operator[] (const size_t index) const;
-    inline int lastIndex() {if (!_path.empty()) return _path.back(); return 0;}
+    IndexPath(const size_t index);
+    size_t& operator[] (const size_t index);
+    const size_t& operator[] (const size_t index) const;
+    inline size_t lastIndex() {if (!_path.empty()) return _path.back(); return 0;}
     const IndexPath operator+(const IndexPath& rhs);
     void increaseValueAtLastIndex();
     void decreaseValueAtLastIndex();
     size_t count();
-    IndexPath appending(int index);
+    IndexPath appending(size_t index);
     IndexPath removingLastIndex();
 
     bool operator<( const IndexPath& otherObject ) const;
     bool operator>( const IndexPath& otherObject ) const;
     bool operator=( const IndexPath& otherObject ) const;
 private:
-    std::vector<int> _path;
+    std::vector<size_t> _path;
     friend QDebug operator<<(QDebug debug, const IndexPath &c);
 };
